test for allergy and non-supported structure question in patient details
